+++ tests/AbstractTestCase.php
@@ -62,6 +62,8 @@ abstract class AbstractTestCase extends WebTestCase
             }
         }';
 
+    public const TEST_AVAILABLE_TAGS = ['tag1', 'tag2', 'tag3'];
+
     public const TEXT_FILE_NAME = 'test.txt';
     public const TEXT_FILE_2_NAME = 'test-updated.txt';
     public const PDF_FILE_NAME = 'test.pdf';